Update run_api.py
in your .env file 

OPENAI_API_KEY=YOURKEY
TWILIO_AUTH_TOKEN=YOUR TOKEN
TWILIO_PHONE_NUMBER=+YOURTWILLIOFROM_NUMBER

Once activated, it initiates a user dialogue. You have the option to enhance it using Google Text-to-Speech for a more lifelike interaction. However, I was disappointed to see that the demo provided by the developer didn't match the content of the GitHub repository. This prompted me to create my own version, which is functional but not perfect.